backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1
pyyaml==6.0.2
insightface==0.7.3
opencv-python==4.10.0.84
onnxruntime
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0
python-jose[cryptography]
passlib[bcrypt]
